backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
qrcode[pil]==7.4.2
Pillow==10.1.0
cryptography==41.0.7
python-dotenv==1.0.0
email-validator==2.1.0
apscheduler==3.10.4
imghr
